feat: refactor cost representation from double to int in currency classes

+++ src/main/java/core/Crafting/Crafting_Action.java
@@ -2,6 +2,6 @@ package core.Crafting;
 
 public interface Crafting_Action {
     Crafting_Item apply(Crafting_Item item); // transforms the item
-    double getCost();                        // relative cost of using this action
+    int getCost();                        // relative cost of using this action
     String getName();                        // name for display
 }
