+++ src/main/java/core/Currency/AnnulmentOrb.java
@@ -10,7 +10,7 @@ import java.util.Random;
 
 public class AnnulmentOrb implements Crafting_Action {
 
-	private double cost = 1.0;
+	private int cost = 1;
 
 	private ModType forcedType = ModType.ANY;
 
@@ -33,7 +33,7 @@ public class AnnulmentOrb implements Crafting_Action {
 	}
 
 	@Override
-	public double getCost() {
+	public int getCost() {
 		return cost;
 	}
